Fix ESLint (#188)
* fix 'lint' npm script and add as CI step

The 'lint' npm script as originally written wasn't working because a path was never specified.  (a file or folder path is required).

Additionally, it was not linting typescript files because .ts and .tsx are not checked by default.  This enables those extensions as well.

Finally, I added a lint step to the GA build

* remove extra eslintrc.json file

While trying to fix issue #175 where VSCode was showing a red squiggly for value TypeScript, I discovered that we have two eslintrc.json files, where in fact we should only need one.  Removing this eslintrc file which specified a babel parser, exposes actual TS lint errors which could not be seen before.

Those errors will be resolved in a future commit.

* Auto fix ESLint errors on save in VSCode

Change the VSCode workspace settings to fix ESLint errors that can be automatically fixed.

* specify eslint env as browser for react

As it turns out, we do need a separate eslintrc.json for the react, and this environment setting seems to be the only thing needed at this point.

See https://stackoverflow.com/questions/41858052/solving-linter-error-no-undef-for-document for more information

* eslint: rename IAnalyticsContext interface

* eslint typescript auto fix errors

The following changes were automatically fixed using eslint --fix

* don't auto fix in npm lint script

I decided against auto fixing linter errors in the npm script as this step is run in GA CI, which is not the place to automatically fix anything.

In VSCode at least, auto fixing linter errors should happen automatically on file save.

* Fix failing CI lint

Co-authored-by: Ira Hopkinson <ira_hopkinson@sil.org>