Fix #212 Error messages not resetting (#214)
* Make error messages reset on click (fix #212)

* Make intro chapter rename ui only

The 0 index chapter was being renamed to "Intro" on import. This was causing an error when trying to render it as there was no folder called "Intro" (see issue #212)

Made the change to intro only occur where the user will see it:
- in the ui for selecting chapters
- in the rendered output file name

* Make getChapterDisplayName util function

Originally put the function in public\sources\util.ts but got the following error:
Relative imports outside of src/ are not supported.