Refactor scripture-app-builder.ts to allow for more timing file options
The way scripture-app-builder.ts was reading the timing files was to simplistic and missing more specific timing info like phrase and word timings.
This is a refactor that takes into account potential blank verse numbers and start and end times being the same.
Preparation for adding in support for phrase and word timings.